Wayfarers of the Wind

+++ Arrays_float/main.cpp
@@ -4,14 +4,14 @@ using namespace std;
 #define tab "\t"
 
 
-void FillRand(double arr[], const int n, int MinRand = 0, int MaxRand = 100);
-void Print(double arr[], const int n);
-void Sum(double arr[], const int n);
-void AVG(double arr[], const int n);
-void MinValueIn(double arr[], const int n);
-void MaxValueIn(double arr[], const int n);
-void Shift_left(double arr[], const int n);
-void Shift_right(double arr[], const int n);
+void FillRand(float arr[], const int n, int MinRand = 0, int MaxRand = 100);
+void Print(float arr[], const int n);
+void Sum(float arr[], const int n);
+void AVG(float arr[], const int n);
+void MinValueIn(float arr[], const int n);
+void MaxValueIn(float arr[], const int n);
+void Shift_left(float arr[], const int n);
+void Shift_right(float arr[], const int n);
 
 void main()
 {
@@ -19,7 +19,7 @@ void main()
 
 	cout << "Алгоритм для вывода суммы массива в double: " << "\n";
 	const int XD = 10;
-	double crr[XD];
+	float crr[XD];
 	FillRand(crr, XD);
 	Print(crr, XD);
 	Sum(crr, XD);
@@ -47,7 +47,7 @@ void main()
 	Print(crr, XD);
 }
 
-void FillRand(double arr[], const int n, int MinRand, int MaxRand)
+void FillRand(float arr[], const int n, int MinRand, int MaxRand)
 {
 	MinRand *= 100;
 	MaxRand *= 100;
@@ -57,7 +57,7 @@ void FillRand(double arr[], const int n, int MinRand, int MaxRand)
 		arr[i] /= 100;
 	}
 }
-void Print(double arr[], const int n)
+void Print(float arr[], const int n)
 {
 	for (int i = 0; i < n; i++)
 	{
@@ -65,27 +65,27 @@ void Print(double arr[], const int n)
 	}
 	cout << endl;
 }
-void Sum(double arr[], const int n)
+void Sum(float arr[], const int n)
 {
-	double sum = 0;
+	float sum = 0;
 	for (int i = 0; i < n; i++)
 	{
 		sum = sum + arr[i];
 	}
 	cout << "\n" << "сумма элементов массива - " << sum << "\n";
 }
-void AVG(double arr[], const int n)
+void AVG(float arr[], const int n)
 {
-	double sum = 0;
+	float sum = 0;
 	for (int i = 0; i < n; i++)
 	{
 		sum = sum + arr[i];
 	}
 	cout << "\n" << "Среднее арифметическое - " << sum / n << "\n";
 }
-void MinValueIn(double arr[], const int n)
+void MinValueIn(float arr[], const int n)
 {
-	double min = arr[0];
+	float min = arr[0];
 	for (int i = 0; i < n; i++)
 	{
 		if (arr[i] < min)
@@ -95,9 +95,9 @@ void MinValueIn(double arr[], const int n)
 	}
 	cout << "\n" << "минимальное значение массива - " << min << endl;
 }
-void MaxValueIn(double arr[], const int n)
+void MaxValueIn(float arr[], const int n)
 {
-	double max = arr[0];
+	float max = arr[0];
 	for (int i = 0; i < n; i++)
 	{
 		if (arr[i] > max)
@@ -107,10 +107,10 @@ void MaxValueIn(double arr[], const int n)
 	}
 	cout << "\n" << "максимальное значение массива - " << max << endl;
 }
-void Shift_left(double arr[], const int n)
+void Shift_left(float arr[], const int n)
 {
 	int number_of_shifts; cout << "\n" << "Введите количество сдвигов: "; cin >> number_of_shifts;
-	double zero;
+	float zero;
 	for (int j = 0; j < number_of_shifts; j++)
 	{
 		zero = arr[0];
@@ -120,12 +120,12 @@ void Shift_left(double arr[], const int n)
 		}
 	}
 }
-void Shift_right(double arr[], const int n)
+void Shift_right(float arr[], const int n)
 {
 	int number_of_shifts; cout << "\n" << "Введите количество сдвигов: "; cin >> number_of_shifts;
 	for (int i = 0; i < number_of_shifts; i++)
 	{
-		double buffer = arr[n - 1];
+		float buffer = arr[n - 1];
 		for (int i = n - 1; i >= 0; i--)
 		{
 			i - 1 >= 0 && i < n ? arr[i] = arr[i - 1] : arr[i] = buffer;
